Back para eliminar usuario terminado
- Se finalizo la funcionalidad para eliminar un usuario correctamente

+++ src/main/java/com/example/app/servicios/UsuarioServicio.java
@@ -62,7 +62,12 @@ public class UsuarioServicio {
         return estaSuscrito;
     }
 
-    public void eliminarCuenta(int id) {
-        UserRepository.deleteById(id);
+    public boolean eliminarCuenta(int id) {
+        if(UserRepository.findById(id).isPresent())
+        {
+            UserRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
